backend/requirements.txt
fastapi==0.111.0
pydantic==2.7.4
uvicorn[standard]==0.30.6
google-generativeai==0.8.3
python-dotenv==1.0.1
requests==2.32.5
